Filtering lists shown on the TabDropdown to the currentUser's lists

+++ src/Components/TabDropdown.js
@@ -118,24 +118,32 @@ export default function TabDropdown(props) {
 	);
 
 	function renderDropdownOptions() {
-		const options = props.lists.map(list => {
-			return (
-				<Option
-					key={list.Id}
-					onClick={() => clickAnOption(list.Id)}
-					selected={list.Id === props.selectedList}
-				>
-					<div>{list.Name}</div>
-					<button
-						onClick={async () => {
-							await props.deleteList(list.Id);
-						}}
-					>
-						&times;
-					</button>
-				</Option>
+		let options = [];
+
+		if (props.currentUser) {
+			const filteredLists = props.lists.filter(
+				list => list.Owner.Id === props.currentUser.Id
 			);
-		});
+
+			options = filteredLists.map(list => {
+				return (
+					<Option
+						key={list.Id}
+						onClick={() => clickAnOption(list.Id)}
+						selected={list.Id === props.selectedList}
+					>
+						<div>{list.Name}</div>
+						<button
+							onClick={async () => {
+								await props.deleteList(list.Id);
+							}}
+						>
+							&times;
+						</button>
+					</Option>
+				);
+			});
+		}
 
 		options.push(
 			<NewListOption key={"add new list"} onClick={clickAddNewList}>
